requirements.txt
flask
flask-cors
langchain-community
langchain-huggingface
faiss-cpu
rapidfuzz
gunicorn
google-cloud-storage